Warning message for ties in the ranked stats
Print a message if there are ties in the metric and the percentage of it.    
This only for metric not equal to zero, as random order of genes equal to 0 it is no relevant.    

No idea which percentage of ties are aceptable, nor if it is relevant for the biological interpretation of the results. But the user should be aware of this when they are doing the analysis. As discussed in #18 and #19