Add option to plotGseaTable to return plot
The previous behavior of plotGseaTable is to call grid.arrange, which renders to the current output device, unlike plotEnrichment which returns a plot. By calling arrangeGrob(), we return an object which can be manually rendered later by the caller by calling grid.draw. This enables the output of plotGseaTable to be stored into a variable and displayed later.